104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,14 +7,14 @@
 
 int main(void)
 {
-	double n1 = 1, n2 = 2, n3, n;
+	long double n1 = 1, n2 = 2, n3, n;
 
-	printf("%.0f, %.0f", n1, n2);
+	printf("%.0Lf, %.0Lf", n1, n2);
 
 	for (n = 3; n <= 98; n++)
 	{
 		n3 = n1 + n2;
-		printf(", %.0f", n3);
+		printf(", %.0Lf", n3);
 		n1 = n2;
 		n2 = n3;
 	}
